+++ src/collective/contact/plonegroup/upgrades/upgrades.py
@@ -2,8 +2,9 @@
 import logging
 from zope.interface import alsoProvides
 from plone import api
+from plone.app.uuid.utils import uuidToObject
 
-from ..config import PLONEGROUP_ORG
+from ..config import FUNCTIONS_REGISTRY, PLONEGROUP_ORG
 from ..interfaces import IPloneGroupContact, INotPloneGroupContact
 
 logger = logging.getLogger('collective.contact.plonegroup: upgrade. ')
@@ -23,3 +24,28 @@ def v2(context):
         obj.reindexObject(idxs='object_provides')
     setup = api.portal.get_tool('portal_setup')
     setup.runImportStepFromProfile('profile-collective.contact.plonegroup:default', 'jsregistry')
+
+
+def v3(context):
+    logger.info("Migrate to v3")
+    portal = api.portal.getSite()
+    pg = portal.portal_groups
+    sg = portal.acl_users.source_groups
+    reg = api.portal.get_registry_record(name=FUNCTIONS_REGISTRY)
+    functions = {}
+    for dic in reg:
+        functions[dic['fct_id']] = dic['fct_title']
+    for group in api.group.get_groups():
+        if '_' not in group.id:
+            continue
+        parts = group.id.split('_')
+        org_uid = parts[0]
+        function = '_'.join(parts[1:])
+        org = uuidToObject(org_uid)
+        if not org or function not in functions:
+            continue
+        full_title = org.get_full_title(separator=' - ', first_index=1)
+        group_title = '%s (%s)' % (full_title.encode('utf8'), functions[function].encode('utf8'))
+        if group.getProperty('title') != group_title or sg._groups[group.id]['title'] != group_title:
+            logger.info("Correcting group %s" % group.id)
+            pg.editGroup(group.id, title=group_title)